test/reftest/vendor/blink/fast/dom/CSSStyleDeclaration/css-style-declaration-named-setter.html

<html>
<head>
<script src="../../../resources/wpt-test-pre.js"></script>
<script src="../../../resources/js-test-pre.js"></script>
</head>
<body style="margin: 0px">
<script>

shouldBeEqualToString("document.body.style.margin", "0px");
shouldBe("document.body.style.margin = 1", "1");

// [comment] does not work on no-quirks-mode
// shouldBeEqualToString("document.body.style.margin", "1px");

// var badString = { toString: function() { throw "Exception in toString()"; } };
var badString = { toString: function() { throw "Exception in toString()"; } };
shouldThrow("document.body.style.margin = badString", "'Exception in toString()'");
// [comment] does not work on no-quirks-mode
// shouldBeEqualToString("document.body.style.margin", "1px"); // Should not reset the previous value.

// [comment] does not work on no-quirks-mode
// // Creating a new property on prototype shadows the named property.
// shouldBe("document.body.style.__proto__.margin = 2", "2");
// shouldBe("document.body.style.__proto__.margin", "2");
// shouldBeEqualToNumber("document.body.style.margin", 2);

// shouldBe("document.body.style.margin = 3", "3");
// shouldBeEqualToNumber("document.body.style.margin", 3);

</script>
<script src="../../../resources/js-test-post.js"></script>
<script src="../../../resources/wpt-test-post.js"></script>
</body>
</html>
